+++ src/window_commands.py
@@ -41,10 +41,21 @@ li ul {
     margin: 2px 0 4px;
 }
 '''
+FRONTMATTER = {
+    "allow_code_wrap": True,
+    "markdown_extensions": [
+        "markdown.extensions.admonition",
+        "markdown.extensions.attr_list",
+        "pymdownx.emoji",
+        "pymdownx.progressbar",
+        "pymdownx.saneheaders",
+        {"pymdownx.smartsymbols": {"ordinal_numbers": False}},
+        "pymdownx.tasklist"
+    ]
+}
 PKG_NAME = __package__.split('.')[0]
 
 
-# TODO: add type hints
 class WkhtmltopdfOpenDocs(sublime_plugin.WindowCommand):
 
     def run(self, resource_path='docs/en/README.md'):
@@ -54,15 +65,13 @@ class WkhtmltopdfOpenDocs(sublime_plugin.WindowCommand):
             import mdpopups
             preview_sheet = mdpopups.new_html_sheet(
                 window=w,
-                # TODO: update for Py3.8 with f-strings
                 name='{}/{}'.format(PKG_NAME, resource_path),
-                contents=sublime.load_resource('Packages/{}/{}'.format(PKG_NAME, resource_path)),
+                contents=mdpopups.format_frontmatter(FRONTMATTER) + sublime.load_resource('Packages/{}/{}'.format(PKG_NAME, resource_path)),
                 md=True,
                 css='{}'.format(CSS)
             )
         except Exception as e:
             print('{}: Exception: {}'.format(PKG_NAME, e))
-            # TODO: print(f'{PKG_NAME}: Exception: {e}')
 
     # def is_enabled(self): return bool